Math Operations for dbMatrix Objects

Implements the Math S4groupGeneric functions for dbMatrix objects. This includes various mathematical operations such as logarithms, exponentials, trigonometric functions, and other transformations.

Usage

# S4 method for class 'dbMatrix'
Math(x)

Arguments

x

A dbMatrix object.

Value

A dbMatrix object with the mathematical operation applied to each element.

Details

This method provides implementations for the following Math functions:

Arithmetic and rounding:

abs(), sign(), sqrt(), ceiling(), floor(), trunc()

Cumulative operations:

cummax(), cummin(), cumprod(), cumsum()

Note: cumprod() is not supported

Logarithmic:

log(), log10(), log2(), log1p()

Note: log1p() is not supported

Trigonometric:

cos(), sin(), tan(), acos(), asin(), atan()

cosh(), sinh(), tanh(), acosh(), asinh(), atanh()

cospi(), sinpi(), tanpi()

Note: acosh() asinh() atanh() are not supported

Exponential:

exp(), expm1()

Note: expm1() is not supported

Special functions:

gamma(), lgamma(), digamma(), trigamma()

Note: digamma() trigamma() are not supported

The function applies the specified mathematical operation to each element of the dbMatrix object.

Examples

mat <- matrix(1, nrow = 3, ncol = 3)
dbmat <- as.dbMatrix(mat)
#> ℹ Creating in-memory database connection...
log(dbmat)
#> 3 x 3  dbMatrix of class "dbDenseMatrix"
#> [[ Colnames: 'col1', 'col2', 'col3' ]]
#>                                   
#> row1 0.0000000 0.0000000 0.0000000
#> row2 0.0000000 0.0000000 0.0000000
#> row3 0.0000000 0.0000000 0.0000000
sqrt(dbmat)
#> 3 x 3  dbMatrix of class "dbDenseMatrix"
#> [[ Colnames: 'col1', 'col2', 'col3' ]]
#>                                   
#> row1 1.0000000 1.0000000 1.0000000
#> row2 1.0000000 1.0000000 1.0000000
#> row3 1.0000000 1.0000000 1.0000000
sin(dbmat)
#> 3 x 3  dbMatrix of class "dbDenseMatrix"
#> [[ Colnames: 'col1', 'col2', 'col3' ]]
#>                                   
#> row1 0.8414710 0.8414710 0.8414710
#> row2 0.8414710 0.8414710 0.8414710
#> row3 0.8414710 0.8414710 0.8414710