Create a sparse or dense dbMatrix object

Create an S4 dbMatrix object in sparse or dense triplet vector format.

Usage

createDBMatrix(
  value,
  class = NULL,
  db_path = ":temp:",
  overwrite = FALSE,
  name = "dbMatrix",
  dims = NULL,
  dim_names = NULL,
  ...
)

Arguments

value

data to be added to the database. See details for supported data types (required)

class

class of the dbMatrix: dbDenseMatrix or dbSparseMatrix (required)

db_path

path to database on disk (relative or absolute) or in memory (":temp:")

overwrite

whether to overwrite if table already exists in database (required)

name

table name to assign within database (optional, default: "dbMatrix")

dims

dimensions of the matrix (optional: [int, int])

dim_names

dimension names of the matrix (optional: list(enum, enum))

...

additional params to pass

Details

This function reads in data into a pre-existing DuckDB database. Based on the name and db_path a lazy connection is then made downstream during dbMatrix initialization.

Supported value data types:

dgCMatrix In-memory sparse matrix from the Matrix package

matrix In-memory dense matrix from base R

.mtx Path to .mtx file (TODO)

.csv Path to .csv file (TODO)

tbl_duckdb_connection Table in DuckDB database in ijx format from existing dbMatrix object. dims and dim_names must be specified if value is tbl_duckdb_connection.

Examples

dgc <- dbMatrix:::sim_dgc()
#> Error in as(data, "dgCMatrix"): no method or default for coercing “matrix” to “dgCMatrix”
dbSparse <- createDBMatrix(value = dgc, db_path = ":temp:",
                           name = "sparse_matrix", class = "dbSparseMatrix",
                           overwrite = TRUE)
#> Error in eval(expr, envir, enclos): object 'dgc' not found
dbSparse
#> Error in eval(expr, envir, enclos): object 'dbSparse' not found