Precompute a dbMatrix table in a database

Precomputes a dbMatrix table in a specificied database connection. This can speed up operations that involve breaking sparsity of a dbSparseMatrix, such as in cases when performing + or - arithmetic operations. Need only be run once.

Usage

precompute(conn, m, n, name)

Arguments

conn

duckdb database connection

m

number of rows of precomputed dbMatrix table

n

number of columns of precomputed dbMatrix table

name

name of the precomputed dbMatrix table to be created

Details

The num_rows and num_cols parameters must exceed the maximum row and column indices of the dbMatrix in order to be used for densifying any dbMatrix. Should these params be less than the maximum row and column indices the precomputed densification process will fail.

In such cases, run this function again with a larger n_rows and num_cols, or to manually remove the precomputed table set options(dbMatrix.precomp = NULL) in the R console.

Examples

con = DBI::dbConnect(duckdb::duckdb(), ":memory:")
precompute(con = con , m = 100, n = 100, name = "precomputed_table")
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("duckdb_connection", package = "duckdb")),     new("standardGeneric", .Data = function (x, ...)     standardGeneric("dbListTables"), generic = structure("dbListTables", package = "dbMatrix"),         package = "dbMatrix", group = list(), valueClass = character(0),         signature = "x", default = NULL, skeleton = (function (x,             ...)         stop(gettextf("invalid call in method dispatch to '%s' (no default method)",             "dbListTables"), domain = NA))(x, ...)), <environment>): unable to find an inherited method for function ‘dbListTables’ for signature ‘"duckdb_connection"’